@@ -1,6 +1,6 @@
 # This migration comes from blacklight (originally 20140320000000)
 # -*- encoding : utf-8 -*-
-class AddPolymorphicTypeToBookmarks < ActiveRecord::Migration
+class AddPolymorphicTypeToBookmarks < ActiveRecord::Migration[4.2]
   def change
     add_column(:bookmarks, :document_type, :string)
     
