@@ -1,4 +1,4 @@
-class AddColumnsToUsers < ActiveRecord::Migration
+class AddColumnsToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :provider, :string
     add_column :users, :uid, :string
