@@ -1,4 +1,4 @@
-class AddGuestsToUsers < ActiveRecord::Migration
+class AddGuestsToUsers < ActiveRecord::Migration[4.2]
   def change
     add_column :users, :guest, :boolean, default: false
   end
