@@ -1,4 +1,4 @@
-class ChangeEmailFieldForUsers < ActiveRecord::Migration
+class ChangeEmailFieldForUsers < ActiveRecord::Migration[4.2]
   def change
   	remove_index :users, :email
   	add_index :users, :email, unique: false
